fix: split eval_composition_poly.rs
