+++ crates/prover/src/core/backend/web/webgpu/constants.rs
@@ -1,0 +1,19 @@
+pub const N_ROWS: u32 = 512;
+pub const N_CONSTRAINTS: u32 = 1144;
+
+pub const N_STATE: u32 = 16;
+pub const N_LOG_INSTANCES_PER_ROW: u32 = 3;
+pub const N_INSTANCES_PER_ROW: u32 = 1 << N_LOG_INSTANCES_PER_ROW;
+pub const N_LANES: u32 = 16;
+pub const N_EXTENDED_ROWS: u32 = N_ROWS * 4;
+pub const N_ORIGINAL_ROWS: u32 = N_ROWS;
+pub const N_COLUMNS: u32 = 1264;
+pub const N_INTERACTION_COLUMNS: u32 = N_INSTANCES_PER_ROW * 4;
+pub const N_WORKGROUPS: u32 = N_EXTENDED_ROWS * N_LANES / THREADS_PER_WORKGROUP;
+pub const N_EXTEND_TRACE_WORKGROUPS: u32 = 256;
+pub const THREADS_PER_WORKGROUP: u32 = 256;
+
+pub const N_LINE_TWIDDLES_SIZE: u32 = N_EXTENDED_ROWS * N_LANES;
+pub const N_LINE_TWIDDLES_FLAT_SIZE: u32 = N_LINE_TWIDDLES_SIZE * 2;
+pub const N_CIRCLE_TWIDDLES_SIZE: u32 = N_LINE_TWIDDLES_SIZE * 2;
+pub const N_ORIGINAL_TRACE_COLUMNS: u32 = N_COLUMNS + N_INTERACTION_COLUMNS;
