+++ crates/prover/src/core/backend/web/webgpu/mod.rs
@@ -1,4 +1,10 @@
-pub mod compute_composition_polynomial_original_trace;
+mod constants;
+pub mod eval_composition_poly;
 pub mod gpu_common;
+pub mod gpu_types;
 pub mod qm31;
+pub mod serialization;
 pub mod utils;
+
+pub use gpu_types::*;
+pub use serialization::ByteSerialize;
